fix compilation bugs
